requirements.txt
psycopg[binary]>=3.1
python-dotenv>=1.0
feedparser>=6.0
rapidfuzz>=3.0
requests>=2.31
PyYAML>=6.0
lark-oapi>=1.5.3
